refactor: 핸들러 메서드 인터페이스 추출
새로운 매퍼를 추가하기 전 공통으로 사용할 메서드를 `HandlerMethod` 인터페이스로 추출함.
